+++ connector/src/main/scala/com/vertica/spark/json/VerticaJsonScan.scala
@@ -24,7 +24,7 @@ import org.apache.spark.sql.types.StructType
 /**
  * We support reading JSON files by re-using Spark's JSON support implemented in [[JsonTable]].
  * */
-class VerticaJsonScan(config: ReadConfig, readConfigSetup: DSConfigSetupInterface[ReadConfig], jsonSupport: VerticaJsonScanSupport) extends Scan with Batch {
+class VerticaJsonScan(config: ReadConfig, readConfigSetup: DSConfigSetupInterface[ReadConfig], jsonSupport: VerticaJsonTableSupport) extends Scan with Batch {
   private val logger = LogProvider.getLogger(classOf[VerticaScan])
 
   private lazy val batch: Batch = {
@@ -35,7 +35,7 @@ class VerticaJsonScan(config: ReadConfig, readConfigSetup: DSConfigSetupInterfac
         case None => ErrorHandling.logAndThrowError(logger, InitialSetupPartitioningError())
         case Some(partitionInfo) =>
           val sparkSession = SparkSession.getActiveSession.getOrElse(ErrorHandling.logAndThrowError(logger, InitialSetupPartitioningError()))
-          jsonSupport.getJsonScan(partitionInfo.rootPath, Some(readSchema()), sparkSession).toBatch
+          jsonSupport.buildScan(partitionInfo.rootPath, Some(readSchema()), sparkSession).toBatch
       }
     }
   }
