+++ connector/src/test/scala/com/vertica/spark/datasource/wrappers/VerticaScanWrapperBuilderTest.scala
@@ -1,0 +1,13 @@
+package com.vertica.spark.datasource.wrappers
+
+import org.apache.spark.sql.connector.read.ScanBuilder
+import org.scalamock.scalatest.MockFactory
+import org.scalatest.flatspec.AnyFlatSpec
+
+class VerticaScanWrapperBuilderTest extends AnyFlatSpec with MockFactory {
+  it should "build VerticaScanWrapper" in {
+    val builder = mock[ScanBuilder]
+    (builder.build _).expects().returning(mock[VerticaScanWrapper])
+    assert(new VerticaScanWrapperBuilder(builder).build().isInstanceOf[VerticaScanWrapper])
+  }
+}
